# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API davidWebProject",
        "description": "Cette API est conçue pour gérer l'intégration de diverses fonctionnalités dans mon projet de freelance. Elle permet la gestion complète du backend, incluant la création, la mise à jour, la suppression et la récupération de données essentielles telles que les témoignages clients, les projets, les utilisateurs, et d'autres éléments de contenu. L'API alimente mon site web en fournissant les données nécessaires tout en facilitant la gestion administrative des opérations. Elle est sécurisée via l'authentification par jeton et est optimisée pour une utilisation par des applications web et mobiles. Idéale pour un usage interne, cette API centralise et simplifie la gestion de mon activité freelance.",
        "contact": {
            "name": "Support",
            "url": "https://www.davidwebprojects.fr",
            "email": "contact@davidwebprojects.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apifreelance.davidwebprojects.fr",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se connecter",
                "operationId": "login",
                "requestBody": {
                    "description": "Informations d'identification nécessaires pour le login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identifiant",
                                    "password"
                                ],
                                "properties": {
                                    "identifiant": {
                                        "type": "string",
                                        "example": "identifiant"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "identifiant": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "urlPictureProfil": {
                                            "type": "string",
                                            "example": "/profile_images/johndoe.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se déconnecter",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}